+++ include/sys/elftypes.h
@@ -14,11 +14,11 @@
 extern "C" {
 #endif
 
-typedef unsigned long	Elf32_Addr;
+typedef unsigned int 	Elf32_Addr;
 typedef unsigned short	Elf32_Half;
-typedef unsigned long	Elf32_Off;
-typedef long			Elf32_Sword;
-typedef unsigned long	Elf32_Word;
+typedef unsigned int 	Elf32_Off;
+typedef int 			Elf32_Sword;
+typedef unsigned int 	Elf32_Word;
 
 #ifdef	__cplusplus
 }
